Undo-did Mahesh's change for PASSED_VALIDATION; Added validation flag check for dirty PTRs trying to go to wip;   DE35189 Ecap should be 0 bug (typo whoops); Some refactoring;
git-tfs-id: [https://tfs-fm-04.devtools.intel.com/tfs/DefaultCollection_TFS04]$/RDMS/MyDeals/MyDeals_Release/UI/main;C1462746

+++ Intel.MyDeals/app/contract/controllers/contract.controller.js
@@ -2827,7 +2827,7 @@
                     }
                 }
             }
-            var rebateVal = (kitRebateTotalVal - parseInt(data[firstTierRowIndex]["ECAP_PRICE_____20_____1"])) // Kit rebate - KIT ECAP (tier of "-1")
+            var rebateVal = (kitRebateTotalVal - parseFloat(data[firstTierRowIndex]["ECAP_PRICE_____20_____1"])) // Kit rebate - KIT ECAP (tier of "-1")
             return kendo.toString(rebateVal, "$#,##0.00;-$#,##0.00");
         }
 
@@ -3668,8 +3668,8 @@
                 },
                 { reload: true });
         }
-        $scope.backToPricingTable = function () {
-            if ($scope.isPtr) return;
+        $scope.backToPricingTable = function () {    
+        	if ($scope.isPtr) return;
 
             if (!$scope._dirty) {
                 $scope.switchingTabs = true;
